lending-product/src/main/java/com/fusionx/lending/product/resources/MasterDefAccountRuleReconResource.java
package com.fusionx.lending.product.resources;

import javax.validation.constraints.NotBlank;
import javax.validation.constraints.Pattern;

import com.fasterxml.jackson.databind.annotation.JsonSerialize;

@JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
public class MasterDefAccountRuleReconResource {
	
	@NotBlank(message = "{common.not-null}")
	@Pattern(regexp = "YES|NO", message = "{common-enable.pattern}")
	private String realTime;
	
	@NotBlank(message = "{common.not-null}")
	@Pattern(regexp = "YES|NO", message = "{common-enable.pattern}")
	private String endOfDay;
	
	@NotBlank(message = "{common.not-null}")
	@Pattern(regexp = "YES|NO", message = "{common-enable.pattern}")
	private String onDemand;

	public String getRealTime() {
		return realTime;
	}

	public void setRealTime(String realTime) {
		this.realTime = realTime;
	}

	public String getEndOfDay() {
		return endOfDay;
	}

	public void setEndOfDay(String endOfDay) {
		this.endOfDay = endOfDay;
	}

	public String getOnDemand() {
		return onDemand;
	}

	public void setOnDemand(String onDemand) {
		this.onDemand = onDemand;
	}
	
	
}
